refactor: change users route to user
