+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './users/users.module';
+import { UserModule } from './user/user.module';
 import { TrackModule } from './track/track.module';
 
 @Module({
-  imports: [UsersModule, TrackModule],
+  imports: [UserModule, TrackModule],
   controllers: [AppController],
   providers: [AppService],
 })
