+++ src/db/db.user.ts
@@ -1,7 +1,7 @@
 import { randomUUID } from 'crypto';
-import { CreateUserDto } from 'src/users/dto/create-user.dto';
-import { UpdateUserDto } from 'src/users/dto/update-user.dto';
-import { User } from 'src/users/entities/user.entity';
+import { CreateUserDto } from 'src/user/dto/create-user.dto';
+import { UpdateUserDto } from 'src/user/dto/update-user.dto';
+import { User } from 'src/user/entities/user.entity';
 import { InMemoryDatabase } from './db.interface';
 
 export class UserDB
