+++ src/user/user.service.ts
@@ -4,7 +4,7 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
-export class UsersService {
+export class UserService {
   constructor(private readonly service: DB) {}
   create(createUserDto: CreateUserDto) {
     return this.service.users.create(createUserDto);
