--- src/users/users.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
-import DB from 'src/db/db';
-
-@Module({
-  controllers: [UsersController],
-  providers: [UsersService, DB],
-})
-export class UsersModule {}
